% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiddich.R
\name{quiddich-package}
\alias{quiddich-package}
\title{QUick IDentification of DIagnostic CHaracters}
\description{
The package \code{QUIDDICH} provides tools for an automated identification of diagnostic molecular characters, i.e. such columns in a given nucleotide or amino acid alignment that allow to distinguish taxa from each other. These characters can be used to complement formal descriptions of taxa, which are often based on morphological and anatomical features. Especially for morphologically cryptic species, this is helpful. \code{QUIDDICH} distinguishes between four different types of diagnostic characters. For more information, see "Kuehn, A.L., Haase, M. 2019. QUIDDICH: QUick IDentification of DIagnostic CHaracters."
}
\author{
A. Luise Kuehn <luise.kuehn@uni-greifswald.de>
}
